streaming-app/requirements.txt
kafka-python==1.4.3
numpy==1.15.4
pandas==0.22.0
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
